requirements.txt
pygame
numpy
pynput
tqdm

pytorch